pyproject.toml
[project]
name = "kaspa-market-maker-bot"
version = "0.1.0"
description = "High-performance Kaspa market making bot for centralized exchanges (MEXC, Gate, KuCoin, etc.)."
authors = [{ name = "Kaspa MM" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.3.35",
    "websockets>=12.0",
    "httpx>=0.27.0",
    "pydantic>=2.8.2",
    "python-dotenv>=1.0.1",
    "structlog>=24.1.0",
    "prometheus-client>=0.20.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.3.0",
    "ruff>=0.6.8",
    "pytest>=8.3.2",
    "pytest-asyncio>=0.23.8",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I"]
target-version = "py311"


